backend/foodgram/requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
attrs==23.1.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==6.0.0
flake8-docstrings==1.6.0
flake8-isort==6.0.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pillow==10.3.0
pluggy==1.3.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.0.1
PyJWT==2.8.0
pytest==6.2.5
pytest-timeout==2.1.0
python-dotenv==0.19.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.0
toml==0.10.2
tornado==6.3.3
typing_extensions==4.11.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
